// project-web-adm-komsco-portal/types/Code.ts
export interface Code {
  first: boolean;
  code: [
    {
      check: boolean;
      cdDtlDesc: string;
      cdDvcd: string;
      cdDvcdDtlDesc: string;
      cdDvcdNm: string;
      dtlCd: string;
      dtlCdNm: string;
      frstRegDttm: string;
      frstRgsrId: string;
      frstRgsrNm: string;
      hisList: string;
      lastChngDttm: string;
      lastEdtrId: string;
      lastEdtrNm: string;
      lnupSer: string;
      uzYn: string;
    }
  ];
}